fixed a few warnings
git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@4411 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d

+++ externals/maxlib/src/beat.c
@@ -131,7 +131,7 @@ static int beat_evaluate(t_beat *x)
 		for(i = 1; i < 6; i++)
 		{
 			sortp = &(theories[BEAT_LONG - i]);
-			sprintf(info, " %4d[%3d]", sortp->theory, sortp->points);
+			sprintf(info, " %4d[%3d]", (int) sortp->theory, (int) sortp->points);
 			strcat(string, info);
 		}
 		sprintf(info, "  %g %g", clock_getlogicaltime(), x->x_beatexpect);
